+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### v2.0.0
+* Rewritten parsing of story component. Supports rendering multiple levels of children.
+* Moved out to separate repo, since the old one is a ghost town.
+
 ### v1.0.0
 
-* First stable release
+* First stable release (created by mthuret)
